refactor: move utils.ts and constants.ts to the ui folder
