+++ src/commands/crawl.ts
@@ -1,3 +1,6 @@
+import type { ProcessingPipeline } from "../core/ProcessingPipeline.js";
+import type { SourceRegistry } from "../core/SourceRegistry.js";
+import type { CrawlOptions } from "../core/types.js";
 import {
 	ERROR_MESSAGES,
 	FIELD_NAMES,
@@ -5,12 +8,9 @@ import {
 	MENU_LABELS,
 	NAV_VALUES,
 	PROMPT_MESSAGES,
-} from "../cli/constants.js";
-import { validatePositiveIntegerOrEmpty } from "../cli/utils.js";
-import type { ProcessingPipeline } from "../core/ProcessingPipeline.js";
-import type { SourceRegistry } from "../core/SourceRegistry.js";
-import type { CrawlOptions } from "../core/types.js";
+} from "../ui/constants.js";
 import { displayResults, showPostCrawlMenuWithFlow } from "../ui/display.js";
+import { validatePositiveIntegerOrEmpty } from "../ui/utils.js";
 
 export async function handleCrawl(
 	sourceRegistry: SourceRegistry,
